refactor all levels to extend from a single class. Also refactor all activities to share helper methods like hideAndroidUI and findViewsByTag
