+++ app/src/main/java/com/example/csci4176_groupproject/BaseActivity.kt
@@ -1,0 +1,40 @@
+package com.example.csci4176_groupproject
+
+import android.os.Build
+import android.view.View
+import android.view.ViewGroup
+import android.view.WindowInsets
+import android.widget.FrameLayout
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity: AppCompatActivity() {
+    var isFullscreen: Boolean = true
+    lateinit var fullscreenContent: FrameLayout
+
+    fun hideAndroidUI() {
+        // Hide UI first
+        supportActionBar?.hide()
+        isFullscreen = false
+
+        if (Build.VERSION.SDK_INT >= 30) {
+            fullscreenContent.windowInsetsController?.hide(WindowInsets.Type.statusBars() or WindowInsets.Type.navigationBars())
+        }
+    }
+
+
+    fun getViewsByTag(root: ViewGroup, tag: String): ArrayList<View> {
+        val views: ArrayList<View> = ArrayList()
+        val childCount = root.childCount
+
+        for (i in 0 until childCount) {
+            val child: View = root.getChildAt(i)
+            if (child is ViewGroup) {
+                views.addAll(getViewsByTag(child, tag))
+            }
+            if (child.tag != null && child.tag.toString().contains(tag)) {
+                views.add(child)
+            }
+        }
+        return views
+    }
+}
